move noopxxx to test_utils
