+++ clash_lib/src/proxy/mod.rs
@@ -248,50 +248,3 @@ pub trait DialWithConnector {
     /// this must be called before the outbound handler is used
     async fn register_connector(&self, _: Arc<dyn RemoteConnector>) {}
 }
-
-#[derive(Debug)]
-pub struct NoopOutboundHandler {
-    pub name: String,
-}
-
-#[async_trait]
-impl DialWithConnector for NoopOutboundHandler {
-    fn support_dialer(&self) -> Option<&str> {
-        None
-    }
-}
-
-#[async_trait]
-impl OutboundHandler for NoopOutboundHandler {
-    fn name(&self) -> &str {
-        &self.name
-    }
-
-    fn proto(&self) -> OutboundType {
-        OutboundType::Direct
-    }
-
-    async fn support_udp(&self) -> bool {
-        false
-    }
-
-    async fn connect_stream(
-        &self,
-        _sess: &Session,
-        _resolver: ThreadSafeDNSResolver,
-    ) -> io::Result<BoxedChainedStream> {
-        Err(io::Error::new(io::ErrorKind::Other, "noop"))
-    }
-
-    async fn connect_datagram(
-        &self,
-        _sess: &Session,
-        _resolver: ThreadSafeDNSResolver,
-    ) -> io::Result<BoxedChainedDatagram> {
-        Err(io::Error::new(io::ErrorKind::Other, "noop"))
-    }
-
-    async fn support_connector(&self) -> ConnectorType {
-        ConnectorType::None
-    }
-}
